add start button - new stage will load previous stored data at first
