+++ app/src/utils/terminal.ts
@@ -46,7 +46,7 @@ class SocketTerminal {
 
   mount(reference: HTMLElement): void {
     this.term.open(reference)
-    this.term.focus()
+    // this.term.focus()
     this.resizeScreen()
   }
